Fixing build warnings

+++ src/interactive_mid_protocols/CommitmentSchemeSimpleHash.cpp
@@ -35,7 +35,7 @@
 void CmtSimpleHashCommitmentMessage::initFromString(const string & s) {
 	auto vec = explode(s, ':');
 	id = stol(vec[0]);
-	for (int i = 2; i < vec.size(); i++) {
+	for (int i = 2; i < (int) vec.size(); i++) {
 		vec[1] += ":" + vec[i];
 	}
 	c = make_shared<vector<byte>>(vec[1].begin(), vec[1].end());
@@ -60,7 +60,7 @@ void CmtSimpleHashDecommitmentMessage::initFromString(const string & s) {
 		vec[1] += ":" + vec[counter++];
 	vector<byte> random(vec[1].begin(), vec[1].end());
 	r = make_shared<ByteArrayRandomValue>(random);
-	for (int i = counter + 1; i < vec.size(); i++)
+	for (int i = counter + 1; i < (int) vec.size(); i++)
 		vec[counter] += ":" + vec[i];
 	vector<byte> tmp(vec[counter].begin(), vec[counter].end());
 	x = make_shared<vector<byte>>(tmp);
