+++ src/interactive_mid_protocols/SigmaProtocolOrMultiple.cpp
@@ -30,7 +30,7 @@
 
 string SigmaOrMultipleCommonInput::toString() {
 	string output = "";
-	for (int i = 0; i < sigmaInputs.size(); i++) {
+	for (int i = 0; i < (int) sigmaInputs.size(); i++) {
 		output += sigmaInputs[i]->toString();
 		output += ":";
 	}
